+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2025.11.26
+
+### Added
+
+- Designed kernel implementation for DanceBuilder component
+
+### Updated
+
+- Changed design to include DanceBuilder1 class implementations.
+
 ## 2025.10.23
 
 ### Added